web/web-2-jwt/src/requirements.txt
Flask
pyjwt
Flask-SQLAlchemy
